requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
lxml==4.8.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqldf==0.4.2
urllib3==1.26.8
pandasql==0.7.3
ipython==7.32.0
jsonpath==0.82
